backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]
langgraph>=0.2.0
langchain>=0.2.0
langchain-groq
langchain-community
python-dotenv
sentence-transformers
faiss-cpu
numpy
pandas
httpx
pydantic>=2.0
